bug fixes
- updated sandbox to new api
- fixed bugs with asserts
- more type warning fixes
- updated todo list
- various minor bug fixes

+++ REngine3/source/re/app/Application.hpp
@@ -27,7 +27,7 @@ namespace re
 		* EXPORTS: none
 		* PURPOSE: Default constructor. Your child class will need to set up the rest of the engine.
 		*/
-		Application(bool enableLogging, bool enableFileLogging, double gravity);
+		Application(bool enableLogging, bool enableFileLogging, float32 gravity);
 
 		/*
 		* IMPORTS: none
