+++ REngine3/source/re/physics/Box2DManager.cpp
@@ -10,8 +10,8 @@
 
 namespace re
 {
-	Box2DManager::Box2DManager(double gravity)
-		:m_world(b2Vec2(0.0, gravity))
+	Box2DManager::Box2DManager(float32 gravity)
+	:m_world(b2Vec2(0.0, gravity))
 	{
 		m_gravity = b2Vec2(0.0, gravity);
 	}
